move vc functions in a dedicated file
